Création Stats Service et controlleur

+++ Backend/database/migrations/2025_02_10_083146_create_cryptocurrencies_table.php
@@ -11,8 +11,8 @@ return new class extends Migration {
             $table->id();
             $table->string('name')->unique();
             $table->string('image')->nullable();
-            $table->float('currentPrice');
-            $table->float('inStock')->nullable();
+            $table->decimal('currentPrice', 10, 3);
+            $table->decimal('inStock', 10, 3)->nullable();
             $table->timestamps();
         });
     }
